@@ -46,12 +46,10 @@
     "publishNpm": "npm publish --access public"
   },
   "dependencies": {
-    "@mdx-js/mdx": "^1.6.22",
-    "@mdx-js/react": "^1.6.22",
-    "@vitejs/plugin-react-refresh": "^1.1.0",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "vite-plugin-mdx": "^3.5.10"
+    "@mdx-js/mdx": "1.6.22",
+    "@mdx-js/react": "1.6.22",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0"
   },
   "devDependencies": {
     "@types/node": "^17.0.23",
@@ -59,6 +57,7 @@
     "@types/react-beautiful-dnd": "^13.1.2",
     "@types/react-dom": "^17.0.10",
     "@vitejs/plugin-react": "^1.0.7",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
     "framer-motion": "^6.2.8",
     "react-beautiful-dnd": "^13.1.0",
     "react-contenteditable": "^3.3.6",
@@ -68,6 +67,7 @@
     "vite": "^2.8.0",
     "vite-dts": "^1.0.4",
     "vite-plugin-dts": "^1.0.4",
+    "vite-plugin-mdx": "^3.5.10",
     "zustand": "^3.7.2"
   },
   "peerDependencies": {}
